requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
pillow>=10.0.0
flask>=2.3.0
opencv-python>=4.8.0
numpy>=1.24.0
accelerate>=0.21.0
requests>=2.31.0
librosa>=0.10.0
scipy>=1.11.0